MDEV-27240 SIGSEGV in ha_spider::store_lock on LOCK TABLE
The use-after-free on ha_spider::partition_handler_share results
in SIGSEGV.

There is not need to allocate partition_handler_share for
non-partitioned tables, and thus we do not do such allocations.
This fixes the bug.

One thing to be noted here is that the patch only covers the
non-partitioned tables.